Change PropertyBinder to use JSON so it works with any property type
